requirements.txt
# requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
flask>=2.0.0
cryptography>=36.0.0
pyjwt>=2.3.0
joblib>=1.1.0
pytest>=6.2.5
gunicorn>=20.1.0
python-dotenv>=0.19.0
Flask-Cors>=3.0.10